requirements.txt
fastapi
uvicorn
requests
asyncpg
python-dotenv
